clox: build and run instruction

+++ clox/chunk.c
@@ -15,7 +15,7 @@ void initChunk(Chunk *chunk)
 void freeChunk(Chunk *chunk)
 {
     FREE_ARRAY(uint8_t, chunk->code, chunk->capacity);
-    FREE_ARRAY(uint8_t, chunk->lines, chunk->capacity);
+    FREE_ARRAY(int, chunk->lines, chunk->capacity);
     freeValueArray(&chunk->constants);
     initChunk(chunk);
 }
@@ -28,7 +28,7 @@ void writeChunk(Chunk *chunk, uint8_t byte, int line)
         chunk->capacity = GROW_CAPACITY(oldCapacity);
         chunk->code = GROW_ARRAY(uint8_t, chunk->code,
                                  oldCapacity, chunk->capacity);
-        chunk->lines = GROW_ARRAY(uint8_t, chunk->lines,
+        chunk->lines = GROW_ARRAY(int, chunk->lines,
                                   oldCapacity, chunk->capacity);
     }
 
@@ -37,6 +37,7 @@ void writeChunk(Chunk *chunk, uint8_t byte, int line)
     chunk->count++;
 }
 
+// Return the index where the constant was appended
 int addConstant(Chunk *chunk, Value value)
 {
     writeValueArray(&chunk->constants, value);
